requirements.txt
# Backend dependencies
fastapi==0.111.1
uvicorn==0.30.3
pdfplumber==0.11.2
google-generativeai==0.7.2
python-dotenv==1.0.1

# Frontend dependencies
streamlit==1.36.0

# Testing dependencies
pytest==8.3.1
requests==2.32.3
